delete list

+++ app/src/main/java/com/marina/shoppinglist/database/MainViewModel.kt
@@ -45,8 +45,9 @@ class MainViewModel(database: MainDatabase) : ViewModel() {
         dao.deleteNote(id)
     }
 
-    fun deleteShopListName(id: Int) = viewModelScope.launch {
-        dao.deleteShopListName(id)
+    fun deleteShopList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+        if (deleteList) dao.deleteShopListName(id)
+        dao.deleteShopItemsByListId(id)
     }
 
     class MainViewModelFactory(val database: MainDatabase) : ViewModelProvider.Factory {
